+++ packages/editor/src/store/utils/test/notice-builder.js
@@ -120,7 +120,7 @@ describe( 'getNotificationArgumentsForSaveFail()', () => {
 			'',
 			[ 'draft', 'publish' ],
 			[
-				'Publishing failed. Something went wrong.',
+				'Publishing failed. We’ll try to save a backup in this browser. Please try publishing again.',
 				defaultExpectedAction,
 			],
 		],
@@ -129,7 +129,7 @@ describe( 'getNotificationArgumentsForSaveFail()', () => {
 			'',
 			[ 'draft', 'private' ],
 			[
-				'Publishing failed. Something went wrong.',
+				'Publishing failed. We’ll try to save a backup in this browser. Please try publishing again.',
 				defaultExpectedAction,
 			],
 		],
@@ -138,7 +138,7 @@ describe( 'getNotificationArgumentsForSaveFail()', () => {
 			'',
 			[ 'draft', 'future' ],
 			[
-				'Scheduling failed. Something went wrong.',
+				'Scheduling failed. We’ll try to save a backup in this browser. Please try scheduling again.',
 				defaultExpectedAction,
 			],
 		],
@@ -146,7 +146,20 @@ describe( 'getNotificationArgumentsForSaveFail()', () => {
 			'when post is published and edits is published',
 			'',
 			[ 'publish', 'publish' ],
-			[ 'Updating failed. Something went wrong.', defaultExpectedAction ],
+			[
+				'Updating failed. We’ll try to save a backup in this browser. Please try updating again.',
+				defaultExpectedAction,
+			],
+		],
+		[
+			'when the save is an autosave',
+			'',
+			[ 'publish', 'publish' ],
+			[
+				'Auto-save failed. We’ll try to save a backup in this browser. You can also save manually.',
+				defaultExpectedAction,
+			],
+			{ isAutosave: true },
 		],
 	].forEach(
 		( [
@@ -154,6 +167,7 @@ describe( 'getNotificationArgumentsForSaveFail()', () => {
 			errorCode,
 			[ postStatus, editsStatus ],
 			expectedValue,
+			options = {},
 		] ) => {
 			// eslint-disable-next-line jest/valid-title
 			it( description, () => {
@@ -165,6 +179,7 @@ describe( 'getNotificationArgumentsForSaveFail()', () => {
 						post,
 						edits,
 						error,
+						options,
 					} )
 				).toEqual( expectedValue );
 			} );