+++ packages/fields/src/fields/pattern-sync-status/index.tsx
@@ -3,7 +3,7 @@
  */
 import type { Field } from '@wordpress/dataviews';
 import { __, _x } from '@wordpress/i18n';
-// @ts-ignore
+// @ts-expect-error `@wordpress/patterns` is not typed yet.
 import { privateApis as patternPrivateApis } from '@wordpress/patterns';
 
 /**