+++ test/e2e/specs/editor/various/a11y.spec.js
@@ -102,7 +102,9 @@ test.describe( 'a11y (@firefox, @webkit)', () => {
 		await pageUtils.pressKeys( 'access+h' );
 
 		// Click a non-focusable element before the first tabbable within the modal.
-		await page.click( 'role=heading[name="Keyboard shortcuts"i]' );
+		await page
+			.getByRole( 'heading', { name: 'Keyboard shortcuts' } )
+			.click();
 
 		await pageUtils.pressKeys( 'shift+Tab' );
 
@@ -132,14 +134,16 @@ test.describe( 'a11y (@firefox, @webkit)', () => {
 		// this behavior.
 
 		// Open the top bar Options menu.
-		await page.click(
-			'role=region[name="Editor top bar"i] >> role=button[name="Options"i]'
-		);
+		await page
+			.getByRole( 'region', { name: 'Editor top bar' } )
+			.getByRole( 'button', { name: 'Options' } )
+			.click();
 
 		// Open the Preferences modal.
-		await page.click(
-			'role=menu[name="Options"i] >> role=menuitem[name="Preferences"i]'
-		);
+		await page
+			.getByRole( 'menu', { name: 'Options' } )
+			.getByRole( 'menuitem', { name: 'Preferences' } )
+			.click();
 
 		const preferencesModal = page.locator(
 			'role=dialog[name="Preferences"i]'
@@ -213,10 +217,9 @@ test.describe( 'a11y (@firefox, @webkit)', () => {
 		// that doesn't exist. The content only shows 'No blocks found' and it's
 		// not scrollable any longer. Check it's not focusable.
 		await clickAndFocusTab( blocksTab );
-		await page.type(
-			'role=searchbox[name="Search for a block"i]',
-			'qwerty'
-		);
+		await page
+			.getByRole( 'searchbox', { name: 'Search for a block' } )
+			.type( 'qwerty' );
 		await clickAndFocusTab( blocksTab );
 		await pageUtils.pressKeys( 'Shift+Tab' );
 		await expect( closeButton ).toBeFocused();